#22 Rename: page->postDetail 이동
